dynamic password size & random number of elements
Before this version we only have 6 digits password. but this time we can set the size of a password as we want. In the previous version we have 2 capital letters, 2 small letter, 2 numbers and 2 punch signs. But now which element will used how many times will be set randomly so that no one cant guess the password pattern.